Automatically update template version in settings.xml
Historically the templates had different versions in settings, but the feature was never fully implemented.
There are only one set of settings with the current version and that's it.
So we update the template version automatically.

In Future we should remove the version intirely, so we don't have this problem any more.